MNT: move Consolidators to a separate submodule
